[Development]Adds test cases for validity FullMark

+++ Testing_project_10/src/main/java/com/mycompany/testing_project_10/Verfication.java
@@ -50,7 +50,7 @@ public boolean isValidSubjectCode(String code) {
 
 
 
-public boolean isValidFullMark(int mark) {
+public boolean isValidFullMark(double mark) {
     return mark == 100;
 }
